id int -> long

+++ src/main/java/com/example/onlab_egyed_gabor/data/repository/CoachRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CoachRepository extends JpaRepository<CoachEntity, Integer> {
+public interface CoachRepository extends JpaRepository<CoachEntity, Long> {
 }
