+++ src/main/java/com/example/onlab_egyed_gabor/data/repository/GameRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GameRepository extends JpaRepository<GameEntity, Integer> {
+public interface GameRepository extends JpaRepository<GameEntity, Long> {
 }
